pyproject.toml
[tool.poetry]
name = "music-genre-classification"
version = "0.1.0"
description = "Music genre classification project"
authors = ["Ilya Vankov <vankov.ia@phystech.edu>"]
readme = "README.md"
packages = [{include = "music_genre_classification"}]

[tool.poetry.dependencies]
python = ">=3.13,<3.14"
torch = "^2.0.0"
pytorch-lightning = "^2.0.0"
torchaudio = "^2.0.0"
librosa = "^0.10.0"
numpy = "^1.24.0"
scikit-learn = "^1.3.0"
hydra-core = "^1.3.0"
hydra-colorlog = "^1.2.0"
mlflow = "^2.8.0"
dvc = "^3.38.0"
pre-commit = "^3.0.0"
click = "^8.1.0"
gdown = "^5.2.0"
matplotlib = "^3.7.0"
tqdm = "^4.65.0"
gitpython = "^3.1.0"
pillow = "^10.0.0"
python-telegram-bot = "^21.6"

[tool.poetry.group.dev.dependencies]
black = "^24.1.1"
isort = "^5.13.2"
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]
